feat(budgets): make the spend ledger exact, like the settled rows feeding it
Settled costs became exact in #661, but the counters the budget gate enforces
against stayed binary floating point, so an exact amount was narrowed to reach
them. Four completions through a live gateway left `users.spend` at
0.6619999999999999 while every row that summed to it was exact to the
micro-dollar, and the error accumulated across every reconcile until the budget
reset, in the value a 403 is decided against.

Retype the eight budget counters to `NUMERIC(18, 6)` using the `UsdCost` type
`usage_logs.cost` already uses: `users.spend` / `reserved`,
`budgets.max_budget`, `scoped_budgets.max_budget` / `current_spend` /
`reserved_spend`, `workspace_budget_defaults.max_budget`, and
`budget_reset_logs.previous_spend`. The scale matches the settled row, so a
counter never rounds an amount it records.

Make the arithmetic Decimal-safe with it: the conditional UPDATE that
pre-debits, the CASE clamps in `_release_reserved` and `_release_expression`,
and the `spend + reserved < max_budget` comparison. A float added to a NUMERIC
column resolves as double precision in PostgreSQL, so `reserve_budget` and
`reconcile_reservation` widen through `models.money.to_usd` at the boundary
rather than letting a float reach the SQL, and the zeroes those expressions are
built from are a module-level `ZERO`.

The reserve-time estimate stays exact rather than narrowing, decided rather than
inherited: on the stream-without-usage path the estimate is what settles the
usage row (`cost_override=reservation.estimate`), so it is accounting after all,
and a hold released at the amount it was taken needs both sides written the same
way.

Bound what an operator can type, since the exact column has a ceiling the float
one did not. A cap above ~$1T came back as a 500 numeric overflow from
PostgreSQL; it is now a 422 off `MAX_USD_LIMIT`, which the OpenAPI schema
advertises, and a deployment that already stored one is told which table and
column before the migration touches any DDL rather than after an ALTER fails at
startup.

The API and dashboard boundaries stay float. Amounts narrow on the way out only,
through `as_float`, and the only OpenAPI change is the new maximum.

Fixes #691

+++ src/gateway/api/routes/embeddings.py
@@ -51,7 +51,7 @@ async def create_embedding(
     # Embeddings have no generated output, so only the prompt contributes cost.
     prompt_chars = sum(len(s) for s in request.input) if isinstance(request.input, list) else len(request.input)
 
-    def estimate(pricing: ModelPricing | None) -> float:
+    def estimate(pricing: ModelPricing | None) -> Decimal:
         return estimate_cost(pricing, prompt_chars=prompt_chars, max_output_tokens=None, default_output_tokens=0)
 
     def usage_tokens(result: CreateEmbeddingResponse) -> tuple[int | None, int | None, int | None]:
